Add files via upload
This is my first-ever coding project on HTML, a cool version of the classic Snake game I made using HTML5 Canvas, Tailwind CSS, and JavaScript! In the game, you move a snake to grab food, and it gets longer and faster each time you score. You have to dodge the walls and not crash into yourself. I added stuff like keyboard and mobile controls, a score tracker, and a high score that saves even if you close the game. The design looks clean with some fun animations. This project was a big challenge for me as a beginner, but it taught me a ton about coding, making games, and building something people can actually play.